+++ AdMob/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 15.0.1.0
+    * This version of the adapters has been certified with AdMob 15.0.1.
+    * Update the native ad adapter to use Google's MediaView and UnifiedNativeAd per requirements (https://developers.google.com/admob/android/native-unified).
+
   * 15.0.0.11
     * Remove manual impression tracking for banner, since AdMob does not have an equivalent callback for impressions (AdListener.onAdImpression() is only applicable for Google's native ads).